【 💡 : refactor】 #1 API構造を整理
